requirements.txt
# Core Flask dependencies
Flask[async]==2.3.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0

# ChatterBot (maintained fork for compatibility)
# This fork is the only way it seems to work with modern Python
git+https://github.com/ShoneGK/ChatterPy

# API and HTTP requests
requests==2.31.0
aiohttp==3.8.6

# Database
SQLAlchemy

# Data processing and analytics
pandas==2.1.1
numpy==1.24.4
scikit-learn==1.3.0
scipy==1.11.3

# Machine Learning
joblib==1.3.2
statsmodels==0.14.0

# Environment and configuration
python-dotenv==1.0.0

# Validation and utilities
pydantic==2.4.2
validators==0.22.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Caching
cachetools==5.3.1

# Date and time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.7

# Logging
structlog==23.1.0

# Rate limiting
slowapi==0.1.9

# Type hints
typing-extensions==4.8.0

# ChatterBot corpus data  
chatterbot-corpus==1.2.0

# spaCy English model (required for ChatterBot)
# Note: Install manually with: python -m spacy download en_core_web_sm
# The model cannot be auto-installed via requirements.txt
# Alternatively just run the setup.py script for automatic install

# Additional dependencies
httpx==0.25.0
pytz==2023.3
newsapi-python==0.2.7
ChatterBot~=1.1.0a7
setuptools~=78.1.0